Simplify implementation of Point and PointContainer
Point is embedded type inside PointContainer.
Duplicate values in each struct is removed and
processing are done using float64.

+++ orunada/tree/point_interface.go
@@ -4,7 +4,7 @@ type PointInterface interface {
 	// Return the total number of dimensions
 	Dim() int
 	// Return the value X_{dim}, dim is started from 0
-	GetValue(dim int) int
+	GetValue(dim int) float64
 	// Return the distance between two points
 	Distance(point PointInterface) float64
 	// Return the distance between the point and the plane X_{dim}=val
